--- autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/Select/GradeDist/GradeDist.css
@@ -1,31 +1,0 @@
-.grades-container {
-    display: flex;
-    align-items: center;
-    flex-grow: 1;
-
-    max-width: 10rem;
-}
-
-.grades-dist {
-    display: flex;
-    width: 100%;
-    height: 0.9rem;
-
-    border-radius: 5px;
-    /* hide the overflow so it actually applies the border-radius on the children */
-    overflow: hidden;
-
-    padding: 0;
-    margin-left: 5px;
-    margin-right: 5px;
-}
-
-.gpa-underline {
-    border-bottom: 1px solid rgb(221, 221, 221);
-}
-
-.no-grades-available {
-    margin-right: 5px;
-    display: flex;
-    align-items: center;
-}
