--- autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/Select/InstructionalMethodIcon/InstructionalMethodIcon.css
@@ -1,4 +1,0 @@
-.instructional-method-container {
-    display: flex;
-    align-items: center;
-}
