--- autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/Select/Select.css.d.ts
@@ -1,91 +1,0 @@
-declare namespace SelectCssNamespace {
-  export interface ISelectCss {
-    "MuiExpansionPanelSummary-content": string;
-    "accordian-details": string;
-    "accordian-root": string;
-    "accordian-summary": string;
-    accordianDetails: string;
-    accordianRoot: string;
-    accordianSummary: string;
-    "add-bottom-space": string;
-    addBottomSpace: string;
-    "button-root": string;
-    buttonRoot: string;
-    "center-progress": string;
-    centerProgress: string;
-    "dense-list-item": string;
-    denseListItem: string;
-    "divider-container": string;
-    dividerContainer: string;
-    "error-text": string;
-    errorText: string;
-    "fit-content": string;
-    fitContent: string;
-    "indent-for-checkbox": string;
-    indentForCheckbox: string;
-    "less-checkbox-padding": string;
-    lessCheckboxPadding: string;
-    "list-subheader-content": string;
-    "list-subheader-dense": string;
-    listSubheaderContent: string;
-    listSubheaderDense: string;
-    "meeting-days": string;
-    "meeting-info-wrapper": string;
-    "meeting-time": string;
-    "meeting-type": string;
-    meetingDays: string;
-    meetingInfoWrapper: string;
-    meetingTime: string;
-    meetingType: string;
-    muiExpansionPanelSummaryContent: string;
-    "my-icon-button": string;
-    "my-list-item-icon": string;
-    myIconButton: string;
-    myListItemIcon: string;
-    "name-honors-icon": string;
-    nameHonorsIcon: string;
-    "no-extra-space": string;
-    "no-grades-available": string;
-    "no-start-padding": string;
-    noExtraSpace: string;
-    noGradesAvailable: string;
-    noStartPadding: string;
-    "placeholder-text": string;
-    placeholderText: string;
-    "section-details-table": string;
-    "section-name-container": string;
-    "section-rows": string;
-    sectionDetailsTable: string;
-    sectionNameContainer: string;
-    sectionRows: string;
-    "sections-wrapper": string;
-    sectionsWrapper: string;
-    "select-menu": string;
-    "select-root": string;
-    selectMenu: string;
-    selectRoot: string;
-    "sort-menu-position": string;
-    "sort-order-button": string;
-    "sort-order-button-icon": string;
-    "sort-order-button-icon-ascending": string;
-    "sort-type-menu-button": string;
-    "sort-type-menu-button-icon": string;
-    sortMenuPosition: string;
-    sortOrderButton: string;
-    sortOrderButtonIcon: string;
-    sortOrderButtonIconAscending: string;
-    sortTypeMenuButton: string;
-    sortTypeMenuButtonIcon: string;
-    "sub-title": string;
-    subTitle: string;
-    "table-container": string;
-    tableContainer: string;
-  }
-}
-
-declare const SelectCssModule: SelectCssNamespace.ISelectCss & {
-  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
-  locals: SelectCssNamespace.ISelectCss;
-};
-
-export = SelectCssModule;
